integrate utils/history.js into RouteWrapper
