+++ src/components/RouteWrapper/RouteWrapper.jsx
@@ -5,7 +5,6 @@ import hoistNonReactStatics from '@helpscout/react-utils/dist/hoistNonReactStati
 import getComponentName from '@helpscout/react-utils/dist/getComponentName'
 import { isModifierKeyPressed } from '../../utilities/keys'
 import { isString } from '../../utilities/is'
-import { createLocation } from '../../utilities/history'
 import WithRouterCheck from '../WithRouterCheck'
 
 function noop() {}
@@ -95,6 +94,29 @@ RouteWrapper.propTypes = {
   replace: PropTypes.bool,
   target: PropTypes.string,
   o: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+}
+
+function sanitizePathName(pathname) {
+  return pathname.replace(/\:(.*)\//g, '')
+}
+
+function generateKey() {
+  return Math.random().toString(36).substr(2, 5)
+}
+
+export function createLocation(path, state, key, location = {}) {
+  const a = document.createElement('a')
+  a.href = path || location.pathname
+  const pathname = sanitizePathName(a.pathname)
+  const locationKey = key || generateKey()
+
+  return {
+    state: state || null,
+    pathname: '/' + pathname.split('/').filter(Boolean).join('/'),
+    search: a.search,
+    hash: a.hash,
+    key: locationKey,
+  }
 }
 
 export default RouteWrapper
