+++ src/components/RouteWrapper/RouteWrapper.test.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { mount, shallow } from 'enzyme'
-import RouteWrapper from './RouteWrapper'
+import RouteWrapper, { createLocation } from './RouteWrapper'
 import { Router } from 'react-router'
 
 let push
@@ -239,5 +239,49 @@ describe('onClick', () => {
     })
 
     expect(fetchSpy).not.toHaveBeenCalled()
+  })
+})
+
+describe('createLocation', () => {
+  test('Can create a history from a plain string path', () => {
+    const location = createLocation('/hello')
+
+    expect(location.pathname).toBe('/hello')
+    expect(location.key).toBeTruthy()
+  })
+
+  test('Strips params from path', () => {
+    const location = createLocation('/hello/:post/:id/there')
+
+    expect(location.pathname).toBe('/hello/there')
+    expect(location.key).toBeTruthy()
+  })
+
+  test('Can customize key for path argument', () => {
+    const key = 'abc123'
+    const location = createLocation('/hello', null, key)
+
+    expect(location.pathname).toBe('/hello')
+    expect(location.key).toBe(key)
+  })
+
+  test('Can create location from another location', () => {
+    const prevLocation = createLocation('/hello')
+    prevLocation.pathname = '/hello/there'
+    const location = createLocation(null, null, null, prevLocation)
+
+    expect(location.pathname).toBe('/hello/there')
+  })
+
+  test('Can customize key for location argument', () => {
+    const prevLocation = createLocation('/hello')
+    prevLocation.pathname = '/hello/there'
+
+    const key = 'abc123'
+
+    const location = createLocation(null, null, key, prevLocation)
+
+    expect(location.pathname).toBe('/hello/there')
+    expect(location.key).toBe(key)
   })
 })
